requirements.txt
# Обязательные зависимости
beautifulsoup4==4.12.2
lxml==6.0.2
openpyxl==3.1.2
click==8.1.7

# Опциональные зависимости (для расширенных функций)
requests==2.31.0

# Телеметрия Monium (OpenTelemetry)
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# Разработка
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
mypy==1.7.0